+++ src/org/helioviewer/jhv/view/j2k/image/ResolutionSet.java
@@ -62,6 +62,7 @@ public class ResolutionSet {
 
         public final int width;
         public final int height;
+        public final SubImage subImage;
 
         public final double factorX;
         public final double factorY;
@@ -70,6 +71,8 @@ public class ResolutionSet {
             level = _level;
             width = _width;
             height = _height;
+            subImage = new SubImage(0, 0, width, height, width, height);
+
             factorX = _factorX;
             factorY = _factorY;
         }